event

+++ EventManagementWebApp/EventManagementAPI/Repositories/Interfaces/IVenueRepository.cs
@@ -4,7 +4,7 @@ namespace EventManagementAPI.Repositories.Interfaces
 {
     public interface IVenueRepository
     {
-        Task<Venue?> GetByIdAsync(int id);
+        Task<Venue?> GetByIdAsync(long id);
         Task<List<Venue>> GetAllAsync();
         Task AddAsync(Venue venue);
         Task UpdateAsync(Venue venue);
